backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.5
python-dotenv>=1.0
drf-spectacular>=0.27
django-filter>=24.2
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
pillow>=12.0
